build: fix project not building on arm macbooks
Turns out arm processors installed in Apple's
macbooks do not support avx2 instructions, yet
Apple Clang builds this project with them enabled
without any warning.

Created executable fails as soon as it reaches
mistakenly inserted avx instruction with illegal
instruction trap.

This commit removes -mavx2 flag as it's not
not needed for this project and was only added
because of copy-pasta from other project that
required them to operate.